dynamic manage cart quantities
